fix(vuex store): cleanup store setup
