+++ src/store/actions.js
@@ -1,5 +1,2 @@
-import { pages } from '../config';
-
 export default {
-
 }
