+++ src/store/modules/current-tournament/index.js
@@ -1,0 +1,13 @@
+import getters from './getters';
+import actions from './actions';
+import mutations from './mutations';
+
+export default {
+  namespaced: true,
+  state: {
+    id: undefined
+  },
+  getters,
+  mutations,
+  actions
+}
